osi: dma: Fix Tx/Rx ring length configuration
Issue: DMA Tx/Rx ring length not updating correctly
after changing with ethtool.

The issue was caused by the ring length being read from
the hardware and then logically OR'd with the new ring length value. This resulted in the ring length being set to the value configured during boot, as the hardware was not reset in this process.
Instead of reading the value from the hardware and logically OR'ing it with the new value, directly set the ring length value to the hardware.

Fix:
Instead of reading the value from the hardware and logically OR'ing
it with the new value, directly set the ring length value to the hardware.

Bug 5158977

Change-Id: I23dc3f18f21c884feea9845e439fbf2783d68bcb
Signed-off-by: Bhadram Varka <vbhadram@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/kernel/nvethernetrm/+/3323922
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: svc-mobile-misra <svc-mobile-misra@nvidia.com>
Reviewed-by: svc-mobile-cert <svc-mobile-cert@nvidia.com>
Reviewed-by: Mahesh Patil <maheshp@nvidia.com>
Reviewed-by: Srinivas Ramachandran <srinivasra@nvidia.com>
Reviewed-by: svcacv <svcacv@nvidia.com>
GVS: buildbot_gerritrpt <buildbot_gerritrpt@nvidia.com>