osi:macsec: Do not create SC as part of init
If  Secure Channel is created in the controller its corresponding Key
table has to be initialized else we see parity errors for the Key table.

This change will Make sure SC is not created as part of initialization.

Manual cherry-pick of below change as there is change is file structure
https://git-master.nvidia.com/r/c/kernel/nvethernetrm/+/3323186

Bug 5145470

Change-Id: I129908efc76d244b6ed5466664c9f56acc6981a6
Signed-off-by: Sanath Kumar Gampa <sgampa@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/kernel/nvethernetrm/+/3334769
Reviewed-by: Srinivas Ramachandran <srinivasra@nvidia.com>
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: svc-mobile-misra <svc-mobile-misra@nvidia.com>
GVS: buildbot_gerritrpt <buildbot_gerritrpt@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: Mahesh Patil <maheshp@nvidia.com>
Reviewed-by: svc-mobile-cert <svc-mobile-cert@nvidia.com>